crée turn et match, pas encore fonctionnel

+++ account.h
@@ -4,7 +4,7 @@
 class Account {
     friend class Bank;
     size_t balance;
-    void add(size_t amount);
+    void add(int amount);
 public:
     Account() : balance(3){};
     const size_t& getSolde() const {return balance;}
